Move generate_preset_pass_manager to generate_preset_pass_manager.py
